refactor  rooms reducer
